struct(T1): extract game unit tests into game/tests/
Move the ~750-line #[cfg(test)] mod tests block out of src/game/mod.rs
into src/game/tests/, split by concern (fame, releases, support, deals,
determinism, save_compat, smoke) with the shared harness in tests/mod.rs.
Pure move: every test fn name is unchanged; the only edit is rewriting
super::sim::seeded_game to crate::game::sim::seeded_game now that the
tests sit one module deeper. mod.rs drops from 1061 to 313 lines.
